Fixed death alert

+++ Assets/KillFloorScript.cs
@@ -4,6 +4,7 @@ using UnityEngine;
 
 public class KillFloorScript : MonoBehaviour
 {
+    bool hit = false; 
     private GameObject FindEntityWithStats(GameObject obj)
     {
         CharacterStats stats = obj.GetComponent<CharacterStats>();
@@ -19,12 +20,13 @@ public class KillFloorScript : MonoBehaviour
 
     private void OnCollisionEnter(Collision other)
     {
-        if (other.gameObject.tag == "Player")
+        if (!hit && other.gameObject.tag == "Player")
         {
             Debug.Log("Player has entered the kill floor");
             GameObject playerObj = FindEntityWithStats(other.gameObject);
             CharacterStats stats = playerObj.GetComponent<CharacterStats>();
             stats.TakeDamage(1000);
+            hit = true;
         }
     }
 }
